Update build command
