+++ .eslintrc.json
@@ -10,7 +10,8 @@
     "react",
     "jsx-a11y",
     "@typescript-eslint",
-    "react-hooks"
+    "react-hooks",
+    "react-refresh"
   ],
   "rules": {
     "react-hooks/exhaustive-deps": "error",
@@ -20,8 +21,8 @@
     "jsx-a11y/click-events-have-key-events": "off",
     "jsx-a11y/no-static-element-interactions":"off",
     "jsx-a11y/no-noninteractive-element-interactions":"off",
-    "jsx-a11y/mouse-events-have-key-events":"off"
-
+    "jsx-a11y/mouse-events-have-key-events":"off",
+    "react-refresh/only-export-components": "warn"
   },
   "env": {
     "browser": true,
