+++ package.json
@@ -8,7 +8,7 @@
     "build:dev": "tsc --watch --noEmit",
     "app:dev": "npm-run-all --parallel build:dev vite",
     "app:dev:browser": "HTTPS=true PORT=3000 TS_NODE_PROJECT=tsconfig.script.json node -r ts-node/register -r tsconfig-paths/register scripts/start.ts",
-    "build": "VITE_APP_VERSION=$(git rev-parse --short HEAD) && tsc --noEmit && NODE_OPTIONS=--max_old_space_size=12384 vite build ",
+    "build": "VITE_APP_VERSION=$(git rev-parse --short HEAD) && tsc --noEmit && NODE_OPTIONS=--max_old_space_size=16384 vite build ",
     "build-storybook": "storybook build",
     "coverage": "react-scripts test --colors --coverage",
     "start": "multiplerun dev",
