+++ src/wallets/local-wallet-dialog/manage.tsx
@@ -1,25 +1,21 @@
 
-import { HelpOutline } from '@mui/icons-material'
 import {
     Box,
     Button,
     Divider,
     Grid,
-    Tooltip,
     Typography,
 } from '@mui/material'
 
-import { ButtonOwnProps } from '@mui/material/Button'
-import { FormikContextType, useFormik } from 'formik'
+import { useFormik } from 'formik'
 import React from 'react'
-import * as yup from 'yup'
 import { TextInput } from '@libs/neumorphism-ui/components/TextInput'
 import { DialogProps } from '@libs/use-dialog'
 import { getMnemonic } from 'wallets/logic/storage'
 import { DepositDialogWithButtons } from '@libs/neumorphism-ui/components/DialogWithButtons'
 import { useConfirm } from '@libs/neumorphism-ui/components/useConfirm'
 import { useAlert } from '@libs/neumorphism-ui/components/useAlert'
-import { copyWords } from './mnemonic'
+import { copyWords, passwordForm, passwordValidationSchema } from './helpers'
 
 export type FormReturn = {
     password: string,
@@ -28,79 +24,6 @@ export type FormReturn = {
     password?: never,
     removeMnemonic: boolean
 } | null;
-
-export const passwordValidationSchema = yup.object({
-    password: yup
-        .string()
-        .test({
-            name: 'Can decode mnemonic',
-            message: 'Wrong password',
-            test: (password) => {
-                try {
-                    if (!password) {
-                        return false
-                    }
-                    const decoded_mnemonic = getMnemonic(password)
-                    return !!decoded_mnemonic
-                } catch (e) {
-                    return false
-                }
-            },
-        }),
-})
-
-export function passwordForm(
-    formik: FormikContextType<{ password: string }>,
-    buttonText: string,
-    buttonColor: ButtonOwnProps["color"],
-) {
-    return (
-        <form onSubmit={formik.handleSubmit} className="form-element" style={{ width: "100%" }} >
-            <Grid container gap={3} sx={{ justifyContent: "center", alignItems: "center" }}>
-                <Grid item xs={12} sm={5}
-                    sx={{
-                        minWidth: "300px"
-                    }}>
-                    <TextInput
-                        label={
-                            <>Password <Tooltip title="This password will never be saved anywhere">
-                                <HelpOutline
-                                    aria-label="help"
-                                    sx={{ fontSize: 15 }}
-                                />
-                            </Tooltip></>
-                        }
-                        id="password"
-                        type="password"
-                        name="password"
-                        value={formik.values.password}
-                        onChange={formik.handleChange}
-                        onBlur={formik.handleBlur}
-                        error={
-                            formik.touched.password &&
-                            Boolean(formik.errors.password)
-                        }
-                        helperText={
-                            formik.touched.password &&
-                            formik.errors.password
-                        }
-                        sx={{ display: "flex", justifyContent: "center" }}
-                    />
-                </Grid>
-                <Grid item>
-                    <Button
-                        variant="contained"
-                        type="submit"
-                        color={buttonColor}
-                        sx={{ maxWidth: "400px" }}
-                    >
-                        {buttonText}
-                    </Button>
-                </Grid>
-            </Grid>
-        </ form>
-    )
-}
